pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sql-repair-env"
version = "0.1.0"
description = "SQL Repair Environment for OpenEnv — given a broken SQL query, fix it"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies (required for server functionality)
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    # sqlite3 is part of the Python standard library — no extra dependency needed
    "huggingface-hub>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point — enables: uv run --project . server
server = "sql_repair_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sql_repair_env", "sql_repair_env.server"]
package-dir = { "sql_repair_env" = ".", "sql_repair_env.server" = "server" }
